Starting defining the plant catalog page and relating widgets/providers
The plant catalog page shows a list of plants available in the application. Each app has some parameters, including a list of tags. The user can select in this page a subset of those tags, and as a consequence only the plants which have all those tags are shown on the screen.